backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain-groq
langchain
yt-dlp
youtube-transcript-api
scikit-learn
